# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NewsBot collector",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/discord/webhooks/by/serverAndChannel": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Returns all the known web hooks based on the Server and Channel given.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fancy Server",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "memes",
                        "name": "channel",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discord/webhooks/new": {
            "post": {
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Creates a new record for a discord web hook to post data to.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Server name",
                        "name": "server",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel name",
                        "name": "channel",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discord/webhooks/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Returns the top 100 entries from the queue to be processed.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Deletes a record by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Updates a valid discord webhook ID based on the body given.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discord/webhooks/{id}/disable": {
            "post": {
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Disables a Webhook from being used.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discord/webhooks/{id}/enable": {
            "post": {
                "tags": [
                    "Config",
                    "Discord",
                    "Webhook"
                ],
                "summary": "Enables a source to continue processing.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    }
}